experiment_neo4j/requirements.txt
psycopg2-binary
neo4j
langchain
sentence-transformers